Turn on the generation of links to the man->HTML gateway.
Provide a function which, given the manual page name and the section,
returns the URL of man->HTML gateway to select the correct page.

Translation teams should adjust this as necessary for their language
specific freebsd.dsl so that the man pages that come up are in the correct
language.