The FreeBSD Core Team has approved a new policy with regards to the status of various architectural platforms and requirements for support and development of those platforms.  This policy distinguishes systems into four tiers:
Tier 1: Fully Supported Architectures
Tier 2: Developmental Architectures
Tier 3: Experimental Architectures
Tier 4: Unsupported Architectures

In addition, the Core Team has also approved a policy regarding how
the status of a hardware architecture may be formally changed.  This
policy is new, and subject to review, comment, and modification as
appropriate.  The details of this policy are implemented in this
change to the FreeBSD Committer's Handbook, which all committers
are encouraged to read and review.

Note: Although sparc64 is currently listed as a 'Developmental
Architecture', it is expected that it will become a 'Fully Supported
Architecture' prior to FreeBSD 5.0.  Likewise, it is likely that
ia64 will become a 'Fully Supported Architecture' prior to the
release.